refactor : before multi module

+++ src/main/kotlin/com/example/demo/foodTruck/adapter/inbound/payload/DetailFoodTruckResponse.kt
@@ -7,6 +7,6 @@ data class DetailFoodTruckResponse(
     val name: String,
     val foodType: FoodType,
     var operatingStatus: Boolean,
-    var avgStarRating: Float,
+    var avgStarRating: Double,
     var reviewCount: Long
 )
